# AVM: Adding bmodexp