## Summary

Adds the new opcode `bmodexp` as described in issue #6139 to support modular exponentiation involving byte strings of up to 4096 bytes. Closes #6139

## Test Plan

- Relevant tests added to `assembler_test.go`, `evalStateful_test.go`, & `eval_test.go`
- Opcode is tested with a range of test vectors with function `TestBytesModExp`, covering panic cases, edge cases, acceptance cases and failure cases. Test vectors were generated manually or with Python.