@giuliop 2024-09-24
`bmodexp` is useful also in different scenarios with smaller inputs so we should not penalize those cost-wise, for instance all ZKP protocols based on elliptic curves use it and they try to operate on the smallest field possible while preserving security for efficiency.

As a concrete example, smart contract verifiers for zk-proofs based on the plonk protocol generated by [AlgoPlonk](https://github.com/giuliop/AlgoPlonk) call a teal [implementation](https://github.com/giuliop/AlgoPlonk/blob/main/verifier/templateBN254.go#L364) of it 5 times, using 32-byte inputs for both curve BN254 and BLS12-381.

Considering that a BN254 verifier consumes ~145,000 opcode budget and a BLS12-381 verifier ~185,000, `bmodexp` would really help bring down that cost.