@giuliop 2024-09-24
> I'm doing this from phone, but `maxStringSize` is the AVM max of 4096? That's unusual for the bmath functions, they normally have a maximum of 128 bytes. I suppose you want more because RSA uses really large keys?

Isn't the maximum [64](https://github.com/algorand/go-algorand/blob/2a02530a933ad2f6bd53eddb967f6e97979d74e6/data/transactions/logic/eval.go#L53) bytes?