@jannotti 2024-09-25
Just to close the loop, is the suggestion that this: `x = exponent_length * max(base_length, modulus_length)**2` will work for the cost function (with appropriate scaling)?

I would be on board with that, and I'll just have to write an "escape hatch" to allow an arbitrary cost function to be written.