@algorandskiy 2024-09-25
Based on this [eval](https://drive.google.com/file/d/1Z6HjmV8scdztfz7ZM6f2VUNRP8nkysYd/view) some log-poly formula gives better approximation.