@giuliop 2024-09-25
Looks reasonable to me, the # of iterations is `exponent_length` and the amount of work performed per iteration is in first approximation modulo(base * base) so proportional to `max(base_length, modulus_length)**2`