@giuliop 2024-09-26
I benchmarked `bmodexp` on my own and  `exponent_length * max(base_length, modulus_length)**2` looks reasonable to me.

I benchmarked using the same byte length for all three inputs, base, mode, and exp and I get:
| Benchmark                                  | Iterations | Time (ns/op)    | Extra (extra/op) |
|--------------------------------------------|------------|-----------------|------------------|
| BenchmarkBModExp/modexp_32_bytes_inputs-12 | 78,057     | 15,019 ns/op    | 4.000 extra/op   |
| BenchmarkBModExp/modexp_64_bytes_inputs-12 | 18,218     | 66,383 ns/op    | 4.000 extra/op   |
| BenchmarkBModExp/modexp_128_bytes_inputs-12| 3,776      | 322,610 ns/op   | 4.000 extra/op   |
| BenchmarkBModExp/modexp_256_bytes_inputs-12| 556        | 2,239,315 ns/op | 4.000 extra/op   |
| BenchmarkBModExp/modexp_512_bytes_inputs-12| 63         | 18,215,421 ns/op| 4.000 extra/op   |
| BenchmarkBModExp/modexp_1024_bytes_inputs-12| 8         | 142,717,656 ns/op| 4.000 extra/op  |
| BenchmarkBModExp/modexp_2048_bytes_inputs-12| 1         | 1,147,616,375 ns/op| 4.000 extra/op |
| BenchmarkBModExp/modexp_4096_bytes_inputs-12| 1         | 9,566,156,542 ns/op| 4.000 extra/op |

If we divide the ns/op by `inputs_length**3` we get:

| Inputs Len | Cost       | Inputs Len³   | Cost / Inputs Len³ |
|-----------|------------|--------------|-------|
| 32        | 15,019     | 32,768       | 0.46   |
| 64        | 66,383     | 262,144      | 0.25   |
| 128       | 322,610    | 2,097,152    | 0.15   |
| 256       | 2,239,315  | 16,777,216   | 0.13   |
| 512       | 18,215,421 | 134,217,728  | 0.14   |
| 1024      | 142,717,656| 1,073,741,824| 0.13   |
| 2048      | 1,147,616,375| 8,589,934,592| 0.13   |
| 4096      | 9,566,156,542| 68,719,476,736| 0.14  |

Looks like the cost stabilizes after inputs length of 128 bytes.
I think we can make the cost proportional to `exponent_length * max(base_length, modulus_length)**2` and perhaps add a constant factor to account for the higher relative cost of small inputs


This is the benchmark function I'm using:
```
func BenchmarkBModExp(b *testing.B) {
	for _, byteLen := range []int{32, 64, 128, 256, 512, 1024, 2048, 4096} {
		var base, exp, mod []byte
		for _, input := range []*[]byte{&base, &exp, &mod} {
			// create random inputs of the given length, we are using math/rand without seeding,
			// so it will generate the same pseudorandom numbers for each run
			*input = make([]byte, byteLen)
			for i := range *input {
				(*input)[i] = byte(rand.Intn(256))
			}
		}
		ops := fmt.Sprintf("byte 0x%x; byte 0x%x; byte 0x%x; bmodexp; pop", base, exp, mod)
		b.Run(fmt.Sprintf("modexp_%d_bytes_inputs", byteLen), func(b *testing.B) {
			benchmarkOperation(b, "", ops, "int 1")
		})
	}
}
```