@algorandskiy 2024-09-26
@mangoplane please commit the benchmarks into `crypto_test.go` - that's where other cost evaluation benchmarks for crypto opcodes live.