@giuliop 2024-09-27
> @giuliop how about re-running modexp_1024+ (like `-count=64 I guess)` to get a better avg value?

sure, I rerun using `-benchtime=64x` to have 64 runs (except for 4096 bytes which would timeout, so I ran it 32 times) and the results are in line as before:

| Benchmark                                          | Trials | Time per op (ns/op) | Extra ops |
| -------------------------------------------------- |:------:|:------------------:|:---------:|
| BenchmarkBModExp1/modexp_1024_bytes_inputs-12      |   64   |   142,368,141       |    4.000  |
| BenchmarkBModExp1/modexp_2048_bytes_inputs-12      |   64   | 1,139,738,240       |    4.000  |
| BenchmarkBModExp1/modexp_4096_bytes_inputs-12      |   32   | 9,443,533,249       |    4.000  |

Looks to me like we can use `exponent_length * max(base_length, modulus_length)**2` and not make it more complicated than that