@algorandskiy 2024-10-03
I reproduced the results and got the log-poly model fits my data points perfectly
```
log_A coefficient: 0.0
log_B coefficient: 0.30924400959422704
log_C coefficient: 1.2483621313283038
log_A_log_B coefficient: 0.0
log_A_log_C coefficient: 0.0
log_B_log_C coefficient: 0.10928549552529052
log_A_log_B_log_C coefficient: 8.62725548783673e-05
Intercept: 0.0
R-squared (on log scale): 0.9964
Root Mean Squared Error: 98143.7211 gas units
```
![image](https://github.com/user-attachments/assets/b4680d6d-6ce7-46db-9f4e-0ded6ffc13e2)