@giuliop 2024-10-06
For the cost function I would use this formula to determine the actual opcode cost, i.e., the number of ticks:

`max(base_lenght, mod_lenght)^1.63 * exp_length / 15 + 200`

where lenght is the # of bytes.

This is how the benchmarking looks like, here I am keeping base and mod at the same byte length and varying the exp length. I am calculating the # of ticks by dividing the ns/op coming from `go test` by 15.

| Base & Mod Len | Exp Len | Average ns/op | Average 'ticks' | Formula 'ticks' |
|----------------|---------|------------|-----------------|-----------------|
| 32             | 16      | 6,445      | 430             | 503             |
| 32             | 24      | 9,274      | 618             | 654             |
| 32             | 32      | 12,665     | 844             | 806             |
| 64             | 16      | 18,432     | 1,229           | 1,138           |
| 64             | 32      | 33,163     | 2,211           | 2,076           |
| 64             | 48      | 52,550     | 3,503           | 3,013           |
| 64             | 64      | 70,248     | 4,683           | 3,951           |
| 128            | 32      | 82,731     | 5,515           | 6,005           |
| 128            | 64      | 160,046    | 10,670          | 11,810          |
| 128            | 96      | 239,196    | 15,946          | 17,615          |
| 128            | 128     | 315,648    | 21,043          | 23,420          |
| 256            | 64      | 579,067    | 38,604          | 36,135          |
| 256            | 128     | 1,166,991  | 77,799          | 72,070          |
| 256            | 192     | 1,652,136  | 110,142         | 108,006         |
| 256            | 256     | 2,121,371  | 141,425         | 143,941         |

This is the function used for the benchmarking:
```
func BenchmarkBModExp2(b *testing.B) {
    // Define the base and mod lengths, and corresponding exp lengths
    for _, byteLen := range []int{32, 64, 128, 256} {
        var expLens []int
        switch byteLen {
        case 32:
            expLens = []int{16, 24, 32}
        case 64:
            expLens = []int{16, 32, 48, 64}
        case 128:
            expLens = []int{32, 64, 96, 128}
        case 256:
            expLens = []int{64, 128, 192, 256}
        }

        for _, expLen := range expLens {
            // Generate base and mod of length byteLen
            base := make([]byte, byteLen)
            mod := make([]byte, byteLen)
            for _, input := range []*[]byte{&base, &mod} {
                for i := range *input {
                    (*input)[i] = byte(mathrand.Intn(256))
                }
            }

            // Generate exp of varying lengths
            exp := make([]byte, expLen)
            for i := range exp {
                exp[i] = byte(mathrand.Intn(256))
            }

            ops := fmt.Sprintf("byte 0x%x; byte 0x%x; byte 0x%x; bmodexp; pop", base, exp, mod)

            b.Run(fmt.Sprintf("modexp_base&mod=%d_bytes_exp=%d_bytes", byteLen, expLen), func(b *testing.B) {
                benchmarkOperation(b, "", ops, "int 1")
            })
        }
    }
}
```